functions/gcs-to-bq/requirements.txt
google-cloud==0.34.0
google-cloud-bigquery==1.26.1

